Refactor existing api + model tests into a central location.
